# Increase by x4.8 performance of ToHexString