--- a/src/Nethermind/Nethermind.Benchmark/Core/FromHexBenchmarks.cs
+++ b/src/Nethermind/Nethermind.Benchmark/Core/FromHexBenchmarks.cs
@@ -26,7 +26,9 @@ namespace Nethermind.Benchmarks.Core
         [Benchmark]
         public byte[] Improved()
         {
+#pragma warning disable CS0612 // Type or member is obsolete
             return Bytes.FromHexStringOld(array);
+#pragma warning restore CS0612 // Type or member is obsolete
         }
     }
 }