… src/Nethermind/Nethermind.Core/Extensions/SpanExtensions.cs
… src/Nethermind/Nethermind.JsonRpc.Benchmark/UInt256ToHexStringBenchmark.cs