## Changes:

Is now x4.8 times faster, and Bytes.ByteArrayToHexViaLookup32Safe, Bytes.ToHexString and SpanExtensions.ToHexString all share the same implementation, rather than having 3 different ones

|   Method | ScenarioIndex |       Mean |     Error |    StdDev | Ratio | RatioSD |
|--------- |-------------- |-----------:|----------:|----------:|------:|--------:|
| Improved |             0 |  45.281 ns | 0.3482 ns | 0.4993 ns |  0.21 |    0.00 |
|  Current |             0 | 215.867 ns | 0.6967 ns | 0.9991 ns |  1.00 |    0.00 |
|          |               |            |           |           |       |         |
| Improved |             1 |   2.206 ns | 0.2604 ns | 0.3897 ns |  0.91 |    0.17 |
|  Current |             1 |   2.417 ns | 0.0265 ns | 0.0388 ns |  1.00 |    0.00 |
|          |               |            |           |           |       |         |
| Improved |             2 |  49.155 ns | 0.4488 ns | 0.6717 ns |  0.23 |    0.00 |
|  Current |             2 | 217.019 ns | 0.8983 ns | 1.3445 ns |  1.00 |    0.00 |
|          |               |            |           |           |       |         |
| Improved |             3 |  47.063 ns | 0.8300 ns | 1.1903 ns |  0.22 |    0.00 |
|  Current |             3 | 217.265 ns | 1.1834 ns | 1.7346 ns |  1.00 |    0.00 |
## Types of changes

What types of changes does your code introduce?
_Put an `x` in the boxes that apply_

- [ ] Bugfix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)
- [ ] Documentation Update
- [ ] Code style update (formatting, renaming)
- [x] Refactoring (no functional changes, no api changes)
- [ ] Build related changes
- [ ] Other (please describe): 

## Testing
**Requires testing**

- [x] Yes
- [ ] No

**In case you checked yes, did you write tests??**

- [x] Yes
- [ ] No

**Comments about testing , should you have some** (optional)

Added some odd/even leading zeros tests

## Further comments (optional)

If this is a relatively large or complex change, kick off the discussion by explaining why you chose the solution you did and what alternatives you considered, etc...