@benaadams 2022-12-14
> Would it make sense to optimize `ByteArrayToHexViaLookup32Safe` in similar way too or it is already optimal? 

Can put in a quick change https://github.com/NethermindEth/nethermind/pull/5011; and do some more fancy stuff with intrinsics also