@benaadams 2022-12-15
Is now x4.8 times faster, and Bytes.ByteArrayToHexViaLookup32Safe, Bytes.ToHexString and SpanExtensions.ToHexString all share the same implementation, rather than having 3 different ones

|   Method | ScenarioIndex |       Mean |     Error |    StdDev | Ratio | RatioSD |
|--------- |-------------- |-----------:|----------:|----------:|------:|--------:|
| Improved |             0 |  45.281 ns | 0.3482 ns | 0.4993 ns |  0.21 |    0.00 |
|  Current |             0 | 215.867 ns | 0.6967 ns | 0.9991 ns |  1.00 |    0.00 |
|          |               |            |           |           |       |         |
| Improved |             1 |   2.206 ns | 0.2604 ns | 0.3897 ns |  0.91 |    0.17 |
|  Current |             1 |   2.417 ns | 0.0265 ns | 0.0388 ns |  1.00 |    0.00 |
|          |               |            |           |           |       |         |
| Improved |             2 |  49.155 ns | 0.4488 ns | 0.6717 ns |  0.23 |    0.00 |
|  Current |             2 | 217.019 ns | 0.8983 ns | 1.3445 ns |  1.00 |    0.00 |
|          |               |            |           |           |       |         |
| Improved |             3 |  47.063 ns | 0.8300 ns | 1.1903 ns |  0.22 |    0.00 |
|  Current |             3 | 217.265 ns | 1.1834 ns | 1.7346 ns |  1.00 |    0.00 |