@benaadams 2022-12-15
SafeLookup is improved x1.12 so it hasn't regressed from sharing the method

|     Method |     Mean |    Error |   StdDev | Ratio | RatioSD |
|----------- |---------:|---------:|---------:|------:|--------:|
|   Improved | 45.74 ns | 0.717 ns | 1.051 ns |  0.89 |    0.03 |
| SafeLookup | 51.40 ns | 0.774 ns | 1.159 ns |  1.00 |    0.04 |