@LukaszRozmej 2022-12-16
.NET 7 is now in master branch