@benaadams 2022-12-18
Added the vectorized path which improves the `ByteArrayToHexToHexBenchmark` by x1.5
```
|     Method |     Mean |    Error |   StdDev | Ratio | RatioSD | Alloc Ratio |
|----------- |---------:|---------:|---------:|------:|--------:|------------:|
|   Improved | 39.96 ns | 1.264 ns | 1.892 ns |  0.89 |    0.07 |        1.00 |
| SafeLookup | 59.80 ns | 0.698 ns | 1.001 ns |  1.33 |    0.06 |        1.00 |
|   HexMateA | 45.22 ns | 1.481 ns | 2.217 ns |  1.00 |    0.00 |        1.00 |
```